Refactoring + pre validator

+++ src/task4/task4_test.go
@@ -2,10 +2,10 @@ package task4
 
 import "testing"
 
-var testCase = []struct{
-	in int
-	want int
-} {
+var testCase = []struct {
+	in   int64
+	want int64
+}{
 	{
 		123457468437,
 		0,
@@ -32,8 +32,8 @@ func TestTask4(t *testing.T) {
 	for _, tc := range testCase {
 		got, _ := Task4(tc.in)
 		if got != tc.want {
-				t.Fatalf("Task4(%d) = %d, want %d",
-					tc.in, got, tc.want)
+			t.Fatalf("Task4(%d) = %d, want %d",
+				tc.in, got, tc.want)
 		}
 	}
 }
